chore: added pure comment to pure exports, renamed to be plural
